+++ app/src/main/java/woowacourse/shopping/presentation/ui/common/uimodel/Operator.kt
@@ -1,0 +1,5 @@
+package woowacourse.shopping.presentation.ui.common.uimodel
+
+enum class Operator {
+    PLUS, MINUS
+}
