--- app/src/main/java/woowacourse/shopping/presentation/ui/home/uiModel/Operator.kt
@@ -1,5 +1,0 @@
-package woowacourse.shopping.presentation.ui.home.uiModel
-
-enum class Operator {
-    PLUS, MINUS
-}
